+++ src/helix/points/delete_custom_reward.rs
@@ -31,13 +31,13 @@
 //!     .broadcaster_id("274637212")
 //!     .id("1234")
 //!     .build();
-//! let response: delete_custom_reward::DeleteCustomReward = client.req_delete(request, &token).await?;
+//! let response: delete_custom_reward::DeleteCustomReward = client.req_delete(request, &token).await?.data;
 //! # Ok(())
 //! # }
 //! ```
 //!
 //! You can also get the [`http::Request`] with [`request.create_request(&token, &client_id)`](helix::RequestDelete::create_request)
-//! and parse the [`http::Response`] with [`DeleteCustomRewardRequest::parse_response(&request.get_uri(), response)`](DeleteCustomRewardRequest::parse_response)
+//! and parse the [`http::Response`] with [`DeleteCustomRewardRequest::parse_response(None, &request.get_uri(), response)`](DeleteCustomRewardRequest::parse_response)
 
 use super::*;
 use helix::RequestDelete;
@@ -65,30 +65,6 @@ pub struct DeleteCustomRewardRequest {
 pub enum DeleteCustomReward {
     /// Reward deleted
     Success,
-    /// Bad Request: Query/Body Parameter missing or invalid
-    BadRequest,
-    /// Unauthenticated: Missing/invalid Token
-    AuthFailed,
-    /// Forbidden: The Custom Reward was created by a different client_id or Channel Points are not available for the broadcaster
-    Forbidden,
-    /// Not Found: The Custom Reward doesn’t exist with the id and broadcaster_id specified
-    NotFound,
-}
-
-impl std::convert::TryFrom<http::StatusCode> for DeleteCustomReward {
-    type Error = std::borrow::Cow<'static, str>;
-
-    fn try_from(s: http::StatusCode) -> Result<Self, Self::Error> {
-        match s {
-            http::StatusCode::NO_CONTENT => Ok(DeleteCustomReward::Success),
-            http::StatusCode::BAD_REQUEST => Ok(DeleteCustomReward::BadRequest),
-            http::StatusCode::UNAUTHORIZED => Ok(DeleteCustomReward::AuthFailed),
-            http::StatusCode::FORBIDDEN => Ok(DeleteCustomReward::Forbidden),
-            http::StatusCode::NOT_FOUND => Ok(DeleteCustomReward::NotFound),
-            // http::StatusCode::INTERNAL_SERVER_ERROR => Ok(DeleteCustomReward::InternalServerError),
-            other => Err(other.canonical_reason().unwrap_or("").into()),
-        }
-    }
 }
 
 impl Request for DeleteCustomRewardRequest {
@@ -100,7 +76,31 @@ impl Request for DeleteCustomRewardRequest {
         &[twitch_oauth2::Scope::ChannelManageRedemptions];
 }
 
-impl RequestDelete for DeleteCustomRewardRequest {}
+impl RequestDelete for DeleteCustomRewardRequest {
+    fn parse_inner_response(
+        request: Option<Self>,
+        uri: &http::Uri,
+        response: &str,
+        status: http::StatusCode,
+    ) -> Result<helix::Response<Self, Self::Response>, helix::HelixRequestDeleteError>
+    where
+        Self: Sized,
+    {
+        match status {
+            http::StatusCode::NO_CONTENT => Ok(helix::Response {
+                data: DeleteCustomReward::Success,
+                pagination: None,
+                request,
+            }),
+            _ => Err(helix::HelixRequestDeleteError::InvalidResponse {
+                reason: "unexpected status",
+                response: response.to_string(),
+                status,
+                uri: uri.clone(),
+            }),
+        }
+    }
+}
 
 #[test]
 fn test_request() {
@@ -121,5 +121,5 @@ fn test_request() {
         "https://api.twitch.tv/helix/channel_points/custom_rewards?broadcaster_id=274637212&id=b045196d-9ce7-4a27-a9b9-279ed341ab28"
     );
 
-    dbg!(DeleteCustomRewardRequest::parse_response(&uri, http_response).unwrap());
+    dbg!(DeleteCustomRewardRequest::parse_response(Some(req), &uri, http_response).unwrap());
 }